api/requirements.txt
fastapi[standard]==0.115.6
motor==3.6.0
python-dotenv==1.0.1
bcrypt==4.2.1
black==24.10.0
isort==5.13.2